fix(events): refactor OpenScanHub events
