+++ packit_service/worker/events/openscanhub/abstract.py
@@ -1,7 +1,6 @@
 # Copyright Contributors to the Packit project.
 # SPDX-License-Identifier: MIT
 
-import enum
 from logging import getLogger
 from typing import Optional
 
@@ -19,7 +18,7 @@ from packit_service.worker.events.event import AbstractResultEvent
 logger = getLogger(__name__)
 
 
-class OpenScanHubTaskAbstractEvent(AbstractResultEvent):
+class OpenScanHubEvent(AbstractResultEvent):
     def __init__(
         self,
         task_id: int,
@@ -70,29 +69,3 @@ class OpenScanHubTaskAbstractEvent(AbstractResultEvent):
     def get_non_serializable_attributes(self):
         # build and scan are not serializable
         return [*super().get_non_serializable_attributes(), "build", "scan"]
-
-
-class OpenScanHubTaskFinishedEvent(OpenScanHubTaskAbstractEvent):
-    class Status(str, enum.Enum):
-        success = "success"
-        cancel = "cancel"
-        interrupt = "interrupt"
-        fail = "fail"
-
-    def __init__(
-        self,
-        status: Status,
-        issues_added_url: str,
-        issues_fixed_url: str,
-        scan_results_url: str,
-        **kwargs,
-    ):
-        super().__init__(**kwargs)
-
-        self.status = status
-        self.issues_added_url = issues_added_url
-        self.issues_fixed_url = issues_fixed_url
-        self.scan_results_url = scan_results_url
-
-
-class OpenScanHubTaskStartedEvent(OpenScanHubTaskAbstractEvent): ...
